+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 2019-12-22
+### Added
+- Clear button for planner/painter (Sets all regions as abandoned)
+
 ## 2019-12-19
 ### Added
 - Shareable link for planner/painter